requirements.txt
mss==9.0.2
numpy==2.1.2
Pillow==9.5.0
requests==2.32.3
pytz==2024.2
